python-etl/requirements.txt

psycopg2-binary